add japanese fonts

+++ install/desktop/fonts.sh
@@ -1,10 +1,17 @@
-mkdir -p ~/.local/share/fonts
+if [ "$OMAKUB_OS_ID" == "arch" ]; then
+  paru -S --noconfirm --needed noto-fonts-cjk ttf-cascadia-mono-nerd
+elif [ "$OMAKUB_OS_ID" == "ubuntu" ]; then
+  sudo apt install -y fonts-noto-cjk
 
-cd /tmp
-wget https://github.com/ryanoasis/nerd-fonts/releases/latest/download/CascadiaMono.zip
-unzip CascadiaMono.zip -d CascadiaFont
-cp CascadiaFont/*.ttf ~/.local/share/fonts
-rm -rf CascadiaMono.zip CascadiaFont
+  mkdir -p ~/.local/share/fonts
 
-fc-cache
-cd -
+  cd /tmp
+  wget https://github.com/ryanoasis/nerd-fonts/releases/latest/download/CascadiaMono.zip
+  unzip CascadiaMono.zip -d CascadiaFont
+  cp CascadiaFont/*.ttf ~/.local/share/fonts
+  rm -rf CascadiaMono.zip CascadiaFont
+
+  fc-cache
+  cd -
+
+fi
